core: drop MicroThreading dep from Stride.Core.Serialization, move ContentManager.LoadAsync/ReloadAsync to extensions in Stride.Engine
